# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bulgarian Voice Coach",
    "description": "Voice-enabled web application for teaching Bulgarian to Slavic speakers with real-time speech recognition, synthesis, and grammar error detection.\n\n**Security Note**: This is a local-first application with no authentication by design. All user progress is stored client-side in localStorage. The API is intentionally open as it runs locally and contains no sensitive data.",
    "contact": {
      "name": "Bulgarian Voice Coach Team",
      "email": "support@example.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8001",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/tts": {
      "get": {
        "tags": [
          "tts"
        ],
        "summary": "Text To Speech",
        "description": "Convert text to speech and stream audio",
        "operationId": "text_to_speech_tts_get",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text"
            }
          },
          {
            "name": "track_timing",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Track Timing"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/content/grammar/{grammar_id}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Grammar",
        "description": "Get specific grammar item by ID with L1-specific contrast notes",
        "operationId": "get_grammar_content_grammar__grammar_id__get",
        "parameters": [
          {
            "name": "grammar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grammar Id"
            }
          },
          {
            "name": "l1",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "L1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/content/drills/{grammar_id}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Drills For Grammar",
        "description": "Get drills for a specific grammar item with L1-specific contrast",
        "operationId": "get_drills_for_grammar_content_drills__grammar_id__get",
        "parameters": [
          {
            "name": "grammar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grammar Id"
            }
          },
          {
            "name": "l1",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "L1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/content/mini-lessons/{lesson_id}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Mini Lesson By Id",
        "description": "Get specific mini-lesson by ID",
        "operationId": "get_mini_lesson_by_id_content_mini_lessons__lesson_id__get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/content/mini-lessons/due": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Get Due Lessons",
        "description": "Get mini-lessons due for review based on user's SRS progress",
        "operationId": "get_due_lessons_content_mini_lessons_due_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProgress"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/content/mini-lessons/for-error/{error_pattern}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Lessons For Error",
        "description": "Get mini-lessons that match a specific error pattern",
        "operationId": "get_lessons_for_error_content_mini_lessons_for_error__error_pattern__get",
        "parameters": [
          {
            "name": "error_pattern",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Error Pattern"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/api/config/l1": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Update L1 Language",
        "description": "Update L1 language preference (session-based, not persistent)",
        "operationId": "update_l1_language_api_config_l1_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/content/analyze": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Analyze Text",
        "description": "Analyze Bulgarian text for grammar errors and generate drills with L1 contrast",
        "operationId": "analyze_text_content_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type or code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response format for API consistency"
      },
...
      "UserProgress": {
        "properties": {
          "lesson_progress": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Lesson Progress"
          }
        },
        "type": "object",
        "required": [
          "lesson_progress"
        ],
        "title": "UserProgress",
        "description": "User progress data for SRS calculation"
      }
    }
  },
...
}